modularization: move component to core
