+++ core/src/main/java/com/my/core/Constants.kt
@@ -1,0 +1,4 @@
+package com.my.core
+
+const val KEY_ID = "id"
+const val KEY_SEARCH = "search"
