+++ core/src/main/java/com/my/core/di/Feature.kt
@@ -1,0 +1,7 @@
+package com.my.core.di
+
+import javax.inject.Scope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class Feature
