+++ core/src/main/java/com/my/core/extensions/RxExtensions.kt
@@ -1,0 +1,19 @@
+package com.my.core.extensions
+
+import io.reactivex.Completable
+import io.reactivex.Flowable
+import io.reactivex.Single
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.schedulers.Schedulers
+
+fun <T> Single<T>.applySchedulers() =
+    this.subscribeOn(Schedulers.io())
+        .observeOn(AndroidSchedulers.mainThread())
+
+fun Completable.applySchedulers() =
+    this.subscribeOn(Schedulers.io())
+        .observeOn(AndroidSchedulers.mainThread())
+
+fun <T> Flowable<T>.applySchedulers(): Flowable<T> =
+    this.subscribeOn(Schedulers.io())
+        .observeOn(AndroidSchedulers.mainThread())
