+++ core/src/main/java/com/my/core/extensions/ViewExtensions.kt
@@ -1,0 +1,15 @@
+package com.my.core.extensions
+
+import android.view.View
+
+fun View.show() {
+    this.visibility = View.VISIBLE
+}
+
+fun View.hide() {
+    this.visibility = View.GONE
+}
+
+fun View.invisible() {
+    this.visibility = View.INVISIBLE
+}
