+++ core/src/main/java/com/my/core/mvp/BaseView.kt
@@ -1,0 +1,3 @@
+package com.my.core.mvp
+
+interface BaseView
