+++ core/src/main/java/com/my/core/mvvm/RxViewModel.kt
@@ -1,0 +1,17 @@
+package com.my.core.mvvm
+
+import androidx.lifecycle.ViewModel
+import io.reactivex.disposables.CompositeDisposable
+import io.reactivex.disposables.Disposable
+
+open class RxViewModel : ViewModel() {
+
+    private val compositeDisposable = CompositeDisposable()
+
+    override fun onCleared() {
+        super.onCleared()
+        compositeDisposable.dispose()
+    }
+
+    protected fun Disposable.addToComposite() = compositeDisposable.add(this)
+}
