+++ core/src/main/res/drawable/ic_delete.xml
