+++ core/src/main/res/drawable/list_background.xml
