+++ core/src/main/res/drawable/search_background.xml
