+++ core/src/main/res/values/colors.xml
