+++ core/src/main/res/values/dimens.xml
