--- shared/components/src/main/java/com/my/resources/converter/ConverterDtoToVo.kt
@@ -1,9 +1,0 @@
-package com.my.resources.converter
-
-interface ConverterDtoToVo<Dto, Vo> {
-    fun toValueObject(dto: Dto): Vo
-
-    fun toValueObject(listOfDto: List<Dto>): List<Vo> {
-        return listOfDto.map { dto -> toValueObject(dto) }
-    }
-}
