--- shared/components/src/main/java/com/my/resources/di/Feature.kt
@@ -1,7 +1,0 @@
-package com.my.resources.di
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.RUNTIME)
-annotation class Feature
