--- shared/components/src/main/java/com/my/resources/extensions/LoadImage.kt
@@ -1,16 +1,0 @@
-package com.my.resources.extensions
-
-import android.widget.ImageView
-import com.squareup.picasso.Picasso
-
-fun ImageView.load(path: String) {
-    Picasso.get()
-        .load(path)
-        .into(this)
-}
-
-fun ImageView.load(resId: Int) {
-    Picasso.get()
-        .load(resId)
-        .into(this)
-}
