--- shared/components/src/main/java/com/my/resources/extensions/ResourcesExtensions.kt
@@ -1,8 +1,0 @@
-package com.my.resources.extensions
-
-import android.content.Context
-
-fun Context.getStringByResName(resName: String): String {
-    val id = this.resources.getIdentifier(resName, "string", this.packageName)
-    return this.resources.getString(id)
-}
