--- shared/components/src/main/java/com/my/resources/extensions/RxExtensions.kt
@@ -1,19 +1,0 @@
-package com.my.resources.extensions
-
-import io.reactivex.Completable
-import io.reactivex.Flowable
-import io.reactivex.Single
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.schedulers.Schedulers
-
-fun <T> Single<T>.applySchedulers() =
-    this.subscribeOn(Schedulers.io())
-        .observeOn(AndroidSchedulers.mainThread())
-
-fun Completable.applySchedulers() =
-    this.subscribeOn(Schedulers.io())
-        .observeOn(AndroidSchedulers.mainThread())
-
-fun <T> Flowable<T>.applySchedulers(): Flowable<T> =
-    this.subscribeOn(Schedulers.io())
-        .observeOn(AndroidSchedulers.mainThread())
