--- shared/components/src/main/java/com/my/resources/extensions/ViewExtensions.kt
@@ -1,15 +1,0 @@
-package com.my.resources.extensions
-
-import android.view.View
-
-fun View.show() {
-    this.visibility = View.VISIBLE
-}
-
-fun View.hide() {
-    this.visibility = View.GONE
-}
-
-fun View.invisible() {
-    this.visibility = View.INVISIBLE
-}
