--- shared/components/src/main/java/com/my/resources/mvp/BaseView.kt
@@ -1,3 +1,0 @@
-package com.my.resources.mvp
-
-interface BaseView
