--- shared/components/src/main/java/com/my/resources/mvvm/RxViewModel.kt
@@ -1,17 +1,0 @@
-package com.my.resources.mvvm
-
-import androidx.lifecycle.ViewModel
-import io.reactivex.disposables.CompositeDisposable
-import io.reactivex.disposables.Disposable
-
-open class RxViewModel : ViewModel() {
-
-    private val compositeDisposable = CompositeDisposable()
-
-    override fun onCleared() {
-        super.onCleared()
-        compositeDisposable.dispose()
-    }
-
-    protected fun Disposable.addToComposite() = compositeDisposable.add(this)
-}
