--- shared/data/remote/src/main/AndroidManifest.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.my.remote">
-
-    <uses-permission android:name="android.permission.INTERNET" />
-
-</manifest>
